+++ package.json
@@ -14,7 +14,6 @@
     "@react-navigation/bottom-tabs": "^5.11.11",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
-    "native-base": "^2.15.2",
     "react": "17.0.1",
     "react-native": "0.64.1",
     "react-native-device-info": "^8.1.3",
