--- src/containers/screens/Profile/constants.js
@@ -1,3 +1,0 @@
-const mainColor = 'gray'
-
-export default mainColor
